# docs/swagger.yaml
basePath: /
definitions:
  models.BalanceSummary:
    properties:
      plans:
        items:
          $ref: '#/definitions/models.PaymentPlan'
        type: array
      wallet:
        $ref: '#/definitions/models.WalletSummary'
    type: object
...
  models.PaymentPlan:
    properties:
      amount:
        type: number
      checkout_mode:
        type: string
      code:
        type: string
      currency:
        type: string
      description:
        type: string
      name:
        type: string
      tokens:
        type: integer
    type: object
...
  models.UserProfileResponse:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      current_period_end:
        type: string
      email:
        type: string
      full_name:
        type: string
      isVerified:
        type: boolean
      last_scan_at:
        type: string
      subscription_plan_id:
        type: string
      subscription_status:
        type: string
      tokens_remaining:
        type: integer
      tokens_used_total:
        type: integer
      updated_at:
        type: string
      wallet_balance:
        type: number
    type: object
  models.UserProfileUpdateRequest:
    properties:
      avatar_url:
        example: https://cdn.example.com/avatar.png
        type: string
      full_name:
        example: Jane Doe
        type: string
    type: object
  models.WalletSummary:
    properties:
      email:
        type: string
      token_rate_per_usd:
        example: 10
        type: integer
      token_value_usd:
        example: 0.1
        type: number
      tokens_remaining:
        type: integer
      tokens_used_total:
        type: integer
      updated_at:
        type: string
      wallet_balance:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ganium.ai
    name: Ganium Support
  description: Ganium AI scam detection, security investigation, scanning, authentication
    and payment API.
  license:
    name: Proprietary
  termsOfService: https://ganium.ai/terms
  title: Ganium API
  version: "1.0"
paths:
  /api/balance:
    get:
      description: Returns the current wallet balance together with the token pricing
        plans. The base plan is $1 for 10 tokens.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BalanceSummary'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get balance and pricing
      tags:
      - Wallet
...
  /api/me:
    delete:
      description: Deletes the authenticated user's account and related scans and
        payment history.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete current user account
      tags:
      - Users
    get:
      description: Returns the authenticated user's profile, token counts, and wallet
        balance.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Updates editable profile fields for the authenticated user.
      parameters:
      - description: Profile update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserProfileUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserProfileResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - Users
...
  /api/wallet:
    get:
      description: Returns the authenticated user's token balance and wallet value.
        This endpoint is read-only.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.WalletSummary'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get wallet summary
      tags:
      - Wallet
...
securityDefinitions:
  BearerAuth:
    description: 'JWT authentication. Enter your token using the format: Bearer <token>.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"